LayoutTests/fast/overflow/replaced-child-100percent-height-inside-fixed-container-with-overflow-auto.html

<html>
<head>
<style type="text/css">
#container {
	width: 200px;
	height: 100px;
	overflow: auto; 
}
#replacedElement {
	width: 200%;
	height: 100%;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="container">
	<img id="replacedElement" style="background:green"/>
</div>
<script>
description('Testcase for bug <a href="https://bugs.webkit.org/show_bug.cgi?id=11355">https://bugs.webkit.org/show_bug.cgi?id=11355</a>. \
When a container having overflow:auto has a horizontal scrollbar, \
the scrollbar is to be placed between the inner border edge and the outer padding edge. \
Thus the content height of a replaced child inside the container must not include the height of the horizontal scollbars.');

debug('The height of the inner element box should be 100% of the containers height minus the height of horizontal scrollbar. \
There should be no vertical scrollable content in the container<br>');
shouldBeTrue('document.getElementById("container").scrollHeight == document.getElementById("container").clientHeight');
debug('Container height = Inner replaced element height + scrollbar height');
shouldBeTrue('document.getElementById("container").offsetHeight > document.getElementById("replacedElement").offsetHeight');
document.body.removeChild(document.getElementById('container'));
</script>
<div id="description"></div>
<div id="console"></div>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
